+++ pkg/pgmodel/end_to_end_tests/migrate_test.go
@@ -1,0 +1,36 @@
+// This file and its contents are licensed under the Apache License 2.0.
+// Please see the included NOTICE for copyright information and
+// LICENSE for a copy of the license.
+package end_to_end_tests
+
+import (
+	"context"
+	"testing"
+
+	"github.com/jackc/pgx/v4/pgxpool"
+)
+
+const (
+	expectedVersion = 1
+)
+
+func TestMigrate(t *testing.T) {
+	if testing.Short() {
+		t.Skip("skipping integration test")
+	}
+	withDB(t, *testDatabase, func(db *pgxpool.Pool, t testing.TB) {
+		var version int64
+		var dirty bool
+		err := db.QueryRow(context.Background(), "SELECT version, dirty FROM prom_schema_migrations").Scan(&version, &dirty)
+		if err != nil {
+			t.Fatal(err)
+		}
+		if version != expectedVersion {
+			t.Errorf("Version unexpected:\ngot\n%d\nwanted\n%d", version, expectedVersion)
+		}
+		if dirty {
+			t.Error("Dirty is true")
+		}
+
+	})
+}
