+++ pkg/pgmodel/end_to_end_tests/sql_bench_test.go
@@ -1,4 +1,4 @@
-package pgmodel
+package end_to_end_tests
 
 import (
 	"context"
